LayoutTests/platform/glib/imported/w3c/web-platform-tests/xhr/access-control-and-redirects-async-same-origin.any-expected.txt

CONSOLE MESSAGE: Cannot use wildcard in Access-Control-Allow-Origin when credentials flag is true.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/access-control-basic-allow-star.py due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.web-platform.test:8800/xhr/resources/access-control-basic-allow-no-credentials.py due to access control checks.

PASS Request without credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=* (with star)
PASS Request with credentials is redirected to a cross-origin response with Access-Control-Allow-Origin=* (with star)
PASS Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin
PASS Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin
PASS Request without credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin (no credentials)
PASS Request with credentials is redirected to a cross-origin response with a specific Access-Control-Allow-Origin (no credentials)
